sql/internal: moving PrintType and tests to internal
