+++ sql/internal/specutil/types.go
@@ -1,4 +1,4 @@
-package sqlspec
+package specutil
 
 import (
 	"errors"
